requirements_backup.txt
pandas>=2.1.0
numpy>=1.20.0
scipy>=1.7.0
tushare>=1.2.89
streamlit>=1.31.0
tensorflow>=2.10.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
pytz>=2021.3
requests>=2.26.0
pandas-ta>=0.3.14b0
plotly>=5.18.0
# Added from example for LLM analysis
crewai>=0.12.2
python-dotenv>=1.0.0
pydantic>=2.5.0
yfinance>=0.2.35
# Added for advanced technical analysis
#TA-Lib
